+++ src/_scss/partials/_inventory.scss
@@ -492,6 +492,13 @@
 			border-radius: 5px 0 0 0;
 		}
 
+		.ammo-switch {
+			height: unset;
+			margin-left: 4px;
+			font-size: 12px;
+			font-family: 'Signika';
+		}
+
 		.item-charges {
 			align-items: center;
 			text-align: center;
