+++ src/_scss/partials/_item-info-card.scss
@@ -121,6 +121,14 @@
       line-height: 1;
     }
 
+    .ammo-switch {
+      height: unset;
+      margin-left: 0;
+      margin-bottom: 4px;
+      font-size: 12px;
+      font-family: 'Signika';
+    }
+
     .info-card-states,
     .info-card-trade,
     .info-card-amount,
