+++ src/css/tidy5e.css
@@ -360,6 +360,14 @@
   line-height: 1;
 }
 
+.tidy5e.sheet.actor #item-info-container-content .info-card .ammo-switch {
+  height: unset;
+  margin-left: 0;
+  margin-bottom: 4px;
+  font-size: 12px;
+  font-family: 'Signika';
+}
+
 .tidy5e.sheet.actor #item-info-container-content .info-card .info-card-states,
 .tidy5e.sheet.actor #item-info-container-content .info-card .info-card-trade,
 .tidy5e.sheet.actor #item-info-container-content .info-card .info-card-amount,
@@ -2827,6 +2835,13 @@
 
 .tidy5e.sheet.actor .items-list .item.expanded .item-name .item-image {
   border-radius: 5px 0 0 0;
+}
+
+.tidy5e.sheet.actor .items-list .item .ammo-switch {
+  height: unset;
+  margin-left: 4px;
+  font-size: 12px;
+  font-family: 'Signika';
 }
 
 .tidy5e.sheet.actor .items-list .item .item-charges {
